Port CDT to xcode.

+++ src/main/aerospike/as_operations.c
@@ -533,7 +533,7 @@ bool as_operations_cdt_op(as_operations *ops, const as_bin_name name, as_cdt_opt
 	}
 
 	as_arraylist list;
-	as_arraylist_inita(&list, n + 1);
+	as_arraylist_inita(&list, (uint32_t)n + 1);
 
 	for (size_t i = 0; i < n; i++) {
 		as_cdt_paramtype type = entry->args[i];
